+++ src/inu/schema/status.py
@@ -5,6 +5,7 @@ from ..error import Malformed
 class Status(Schema):
     enabled: bool = None
     active: bool = None
+    locked: bool = None
     status: str = None
 
     def _validate(self):
@@ -15,4 +16,15 @@ class Status(Schema):
             raise Malformed("'active' cannot be None")
 
     def __repr__(self):
-        return f"enabled={self.enabled} active={self.active} status=\"{self.status}\""
+        return f"enabled={self.enabled} locked={self.locked} active={self.active} status=\"{self.status}\""
+
+    def can_act(self, allow_active=False):
+        """
+        Check if the state is suitable for a device to act (eg. on a trigger)
+
+        Returns True if the device is enabled, not locked and not active.
+        """
+        if not allow_active and self.active:
+            return False
+
+        return self.enabled and not self.locked